requirements.txt
# Core ML
torch==2.1.0
transformers==4.40.2
accelerate==0.30.1
bitsandbytes==0.43.1
sentence-transformers==2.7.0
datasets==2.19.1

# Retrieval
rank-bm25==0.2.2
pytrec-eval-terrier==0.5.6

# Evaluation metrics
evaluate==0.4.2
rouge-score==0.1.2
bert-score==0.3.13
nltk==3.8.1

# NLI / Faithfulness
unbabel-comet==2.2.2

# NLP utilities
spacy==3.7.4

# Data handling
pandas==2.1.1
numpy==1.24.3
scipy==1.11.3
scikit-learn==1.3.2
jsonlines==4.0.0

# Config & CLI
pyyaml==6.0.1
python-dotenv==1.0.1

# Visualization
matplotlib==3.8.0
seaborn==0.13.2
tabulate==0.9.0

# Utilities
tqdm==4.66.2
